Working on The Character Movement Stats and overall Controller
Added Scriptable Object for movement stats, Refactored the name of CharacterMovementStats to CharacterMovementSettings to align better with naming convention.

Added custom editor drawer to expand the SO fields inline when it is a reference field in another script. (Allowing the old way of editing SO values at runtime without needing to find the asset file in the project hierarchy.)